+++ app/src/main/java/com/weather/paris/data/network/weather/entity/WeatherInfoEntity.kt
@@ -51,7 +51,7 @@ data class WeatherInfoEntity (
 
     // Midday UV index
     @SerializedName("uvi")
-    var uvIndex: Int,
+    var uvIndex: Float,
 
     // Probability of precipitation
     @SerializedName("pop")
